gen ai legal ass/requirements.txt
streamlit
spacy
python-docx
PyPDF2
openai
fpdf

